requirements.txt
fastapi
uvicorn
sqlmodel
SQLAlchemy
alembic
pydantic
pydantic-settings
python-dotenv
loguru
pwdlib[argon2,bcrypt]
python-jose
psycopg2-binary
python-multipart

